backend/requirements.txt
# Core dependencies
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# LangChain & RAG - Specific versions for stability
langchain-core==0.1.52
langchain==0.1.17
langchain-community==0.0.36
langchain-openai==0.0.5
langchain-huggingface==0.0.1
langchain-chroma==0.1.4
langchain-experimental==0.0.58
chromadb==0.4.24
langsmith>=0.1.0,<0.2.0
tiktoken>=0.5.2,<0.6.0

# Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
presidio-analyzer>=2.2.0,<3.0.0
presidio-anonymizer>=2.2.0,<3.0.0
langdetect==1.0.9

# Database & Cache
motor>=3.3.0,<4.0.0
pymongo>=4.5.0,<5.0.0
redis>=5.0.1,<6.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Monitoring & Observability
prometheus-client>=0.17.1,<1.0.0
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.41b0,<1.0.0

# Performance
ujson>=5.8.0,<6.0.0
orjson>=3.9.10,<4.0.0
aiofiles>=23.2.1,<24.0.0
httpx>=0.25.1,<1.0.0

# Development & Testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<1.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.10.1,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.6.1,<2.0.0